Hiding the IE clear button
As far as I can tell, it doesn't seem possible to target the click on the IE "x" and when it is clicked, the field value is not cleared due to React controlling it. So it does not seem possible to run any functionality when the user tries to clear the value.

Since that is the case, I'm just hiding the clear button instead so that users don't try to click on it.